DIV-1101 Allow sitemap and graph to check nextStep
The Step class throws errors when subclasses fail to overwrite certain
getters. Not every step overides all getters so when they are asserted
against (such as when the graph attempts to determine their type) errors
are thrown within the app and result in a redirect to '/generic-error'.

Return null from getters rather than throw an error. Add test for
siteGraph.